DLF Garden City Enclave Sector 93 Gurgaon
DLF Garden City Enclave Gurugram is an upcoming residential property which is launched by the most reputed developer named DLF India, and it is located at one of the places in Sector 93, Gurgaon. This premium project is now presenting you with Independent Floors and Plots including top-notch amenities and facilities, and all units are amazing and permit much natural light and cross ventilation. Let the positivity discharge all around, in and out to restore your mind, soul, and body behind an active day at work.

DLF Garden City Enclave is designed by renowned architect DLF India and comes in an attractive price bracket that is easy on the pocket. It is a secure place with no vehicular movement and stone-paved arcades connecting all the towers to and from for hassle-free pedestrian access. It is an opulent residential community combined with the amenities of modern living and the beauty of the peaceful countryside.

Complemented with a central water spine, breathtaking multilevel landscaping, and a European-style social club the project. All floors are well-ventilated and have drop-off zones in the basement, resort pool, and active club with a temperature-controlled lap pool, a well-designed service core, and a separate service ramp.

Multiple connectivity routes to important destinations in the Haryana region and established neighborhoods make it an ideal location. Manesar's distance is 10 km, Medanta The Medicity is 16 Km, IFFCO Chowk Metro Station is 23 Km, and Delhi International Airport is 30 km from this place.

DLF Garden City Enclave in Sector 93 is located in Gurgaon. It is conveniently connected to an affluent neighborhood of a millennium city. You can now enjoy relaxed and comfortable living at this exquisite address. It is in close proximity to a metro station, bus stands, hospitals, schools, markets, and IT parks. NH-48 is well connected to this location and makes traveling to major destinations easy. Thus, this property in Haryana is winning the hearts of home seekers as it offers them the dream life. So, book today and head to your customized destination.

For More Info
Visit: - https://bit.ly/3hNWTNL